ANDROID: arm64: mm: ensure that memstart_addr and physvirt_offset remain in sync
In arm64_memblock_init(), physvirt_offset is defined in terms of
PHYS_OFFSET, which is simply an alias for memstart_addr wrapped in a
VM_BUG() check.

Unfortunately, the assignment of physvirt_offset takes place before the
point where the placement of DRAM in the linear region is randomized on
KASLR kernels, which causes memstart_addr to be modified. This leads to
a disparity between the two quantities.

So defer the assignment of physvirt_offset and vmemmap to the point
where memstart_addr has assumed its final value.

Bug: 161946584
Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Cc: Will Deacon <willdeacon@google.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
Change-Id: Ie0e70d825fbd23691a7f5719652b924aaeea29c6