ANDROID: GKI: Update the .xml file after android11-5.4 merge
The .xml file needs to be updated due to 296f9baac0f4 ("ANDROID: GKI:
api preservation of struct inet_connection_sock") in the
android11-5.4-lts branch, which causes a signature difference for two
symbols.

See 296f9baac0f4 ("ANDROID: GKI: api preservation of struct
inet_connection_sock") for all of the details, this is an acceptable
change to make.

Bug: 161946584
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
Change-Id: Ib65395fdc0d0148d310519bf8b6bf1600c81c678