Revert "media: v4l2-core: correctly validate video and metadata ioctls"
This breaks 4k 60fps recording on r9s (specifically with the Samsung Camera app) because the MFC and Camera drivers are set to VFL_TYPE_GRABBER, which was also deprecated in 7c850cb4ffeb1a8173d9476ea6631a19563b91c8.
Changing their types to VFL_TYPE_VIDEO did not fix the issue for an unknown reason, so we can instead revert this change to allow the ioctls to work anyway.

This reverts commit 6714d39f23ffe7fcd8dce980cf573779175d5994.

Signed-off-by: Nahuel Gómez <nahuelgomez329@gmail.com>